Create Lab2.java
Design and implement a CommissionCalculator class that contains (or “has a”) single Auto, Home, and Life insurance policy object (from the classes developed in the previous lab). Create each object in the CommissionCalculator constructor. Also in the CommissionCalculator class, write a method named Run that creates and uses the menu below to prompt the user to enter information for each Auto, Home, and Life policy object and computes commission and prints the commission for each policy object as demonstrated in the example program output below.

Write a driver class as shown below that simply creates a CommissionCalculator object and calls its Run method. Your submission must contain a total of 5 separate class files (Auto, Home, Life, CommissionCalculator, and Driver) to receive full credit.