Create Lab3.java
Design an abstract, base insurance policy class, named Policy, to serve as the parent class for the Auto, Home, and Life insurance policy classes. When designing the Policy class, include the variables and methods that are common to all of the Auto, Home, and Life classes. Write an empty, abstract method in the Policy class named computeCommission for each derived class to implement in order to compute the specific type of commission in their respective class.

 

Re-design the Auto, Home, and Life insurance policy classes from the previous lab as derived classes from this new abstract, base Policy class. Remember to remove variables or methods in each class that are now included in the parent Policy class. Also, implement the abstract computeCommission method inherited from the Policy class in each Auto, Home, and Life class in order to define how to compute the commission for each class.

 

Use your existing CommissionCalculator class and driver class from the previous lab. You do not need to modify the CommissionCalculator and driver classes from the previous lab. Feel free to use classes from the instructor’s solution from the previous lab. This lab should execute exactly the same as the previous lab.

 

 

Hints

Refer to the instructor programs in the Content area under sections Ch9.1-9.2 for examples of how to derive classes from a base (parent) class and sections Ch9.3-9.5 for examples of how to create an abstract method in a parent class and implement the abstract method in derived classes.

Book: Java Software Solutions, 9/e