Create Lab4.java
Re-design the CommissionCalculator class from the previous lab to store an ArrayList of base Policy class object reference variables as shown below to replace the single Auto, Home, and Life policy as in the previous lab.

 
       ArrayList<Policy> policies;

      
Use the menu below to prompt a user to input any number of different types of insurance policies. For options 1-3, use the input information to create the specific type of policy and store it in the newly added ArrayList of base Policy objects above.

After policies have been created and stored, the user can then select option 4 below. This option must use the single loop shown below to compute the commission and print the information for all the policies created stored in the ArrayList of base Policy class objects.

In this way, your program demonstrates polymorphism when calling the base Policy class computeCommisson and toString methods as shown. Continue the program until the user quits. Use all of the other classes (including the driver) from the previous lab. Note you do not need to modify these classes. Feel free to use classes from the instructor’s solution the previous lab.

Refer to the instructor programs in the Content area under sections Ch10.1-10.2 demonstrating the concept of dynamic binding and how we can use base class object reference variables to store references to derived class objects to demonstrate polymorphism.