Create Lab5.java
Design a Java interface named Priority that includes two abstract methods named setPriority and getPriority. Classes that implement this interface use these methods to establish a numeric value (between 1 and 5) that represents priority. Design a class named Task that implements this Priority interface to establish priority among a set of Task objects. Your Task class should store a String to represent the task and an integer representing the priority. Write a driver named TaskList that creates and prints a number of Task objects each with different priorities. Example output of such a list is shown below.


Hints

To assist with the design of this lab, I highly recommend looking closely at the textbook Ch7.5 examples (this lab is very similar to these examples) as well as my examples in the Content area for sections 7.5 on our Cobra home page.