Create Lab6.java
Demonstrate polymorphism with interfaces by developing an application that prompts a user to create five students by entering a first name, last name, and a single test score between 0 and 100. After the students have been entered, sort the students by test score using the selection sort method in descending order (highest to lowest) and print the list of students to the Console. Referencing the examples in the Content area as well as the example in section 10.4 of the textbook:


-        Develop a Student class that implements the Comparable interface.
-        Use (and modify) the Sorting class from section 10.4 to sort the students. Note only a very slight modification is necessary to the Sorting class.
-        In addition to the Student class, the Sorting class, and a driver class, develop a Gradebook class to manage the list of students
-        Use the Gradebook class to store an array of Student objects (NOTE: to work correctly with the Sorting algorithm, do not use an ArrayList)

 
When writing the Gradebook class, be sure to use good design principles. Specifically, think about the information a class should store or “know” (which become the class variables) and the actions a class should be able to perform or “do” (which become the class methods). Be sure to reference the instructor programming examples.